+++ src/main/java/pl/plajer/buildbattle/ConfigPreferences.java
@@ -150,7 +150,7 @@ public class ConfigPreferences {
     BOSSBAR_ENABLED("Boss-Bar-Enabled", true), BUNGEE_ENABLED("BungeeActivated", false), DATABASE_ENABLED("DatabaseActivated", false),
     INVENTORY_MANAGER_ENABLED("InventoryManager", true), WIN_COMMANDS_ENABLED("Win-Commands-Activated", false),
     SECOND_PLACE_COMMANDS_ENABLED("Second-Place-Commands-Activated", false), THIRD_PLACE_COMMANDS_ENABLED("Third-Place-Commands-Activated", false),
-    END_GAME_COMMANDS_ENABLED("End-Game-Commands-Activated", false);
+    END_GAME_COMMANDS_ENABLED("End-Game-Commands-Activated", false), BLOCK_COMMANDS_IN_GAME("Block-Commands-In-Game", true);
 
     private String path;
     private boolean def;